+++ src/main/kotlin/com/codescene/jetbrains/listeners/CodeSceneDynamicPluginListener.kt
@@ -1,16 +1,14 @@
 package com.codescene.jetbrains.listeners
 
 import com.codescene.jetbrains.services.telemetry.TelemetryService
-import com.codescene.jetbrains.util.Constants
+import com.codescene.jetbrains.util.TelemetryEvents
 import com.intellij.ide.plugins.DynamicPluginListener
 import com.intellij.ide.plugins.IdeaPluginDescriptor
 
 class CodeSceneDynamicPluginListener : DynamicPluginListener {
 
     override fun pluginLoaded(pluginDescriptor: IdeaPluginDescriptor) {
-        TelemetryService.Companion.getInstance().logUsage(
-            "${Constants.TELEMETRY_EDITOR_TYPE}/${Constants.TELEMETRY_ON_ACTIVATE_EXTENSION}"
-        )
+        TelemetryService.Companion.getInstance().logUsage(TelemetryEvents.ON_ACTIVATE_EXTENSION)
     }
 
 }