+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Telemetry consent prompt and state (false by default).
+
+### Changed
+- Telemetry events will only be logged once user consent is given.
 
 ## [0.0.9-beta] - 2024-01-29
 ### Added