+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.10-beta] - 2024-01-31
+### Changed
+- Telemetry refactored and improved
+
 ## [0.0.9-beta] - 2024-01-29
 ### Added
 - CodeScene telemetry to track user activity